feat : 즉시결제 구현..

+++ src/main/java/com/ecommerce/chatdemo/domain/member/dto/MyInfoResponse.java
@@ -7,7 +7,7 @@ public record MyInfoResponse(
         String email,
         String name,
         String role,
-        BigDecimal pointBalance,
+        Long pointBalance,
         String membershipName
 ) {
 }
